projeto/SacolaPersonalizacao/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
pillow==10.1.0
python-dotenv==1.0.0
aiofiles==23.2.1
jinja2==3.1.2
email-validator==2.1.0
